linked list only uses key to find elements
previously both key and value of a node needed to match to get
a match. Now only the key is used to find items.

In the hashMap, inside a bucket, in a LL, if the supplied key is
found, that node's value attribute is updated.
Other a new node with the supplied key, value is created